project

+++ project/app/code/local/Core/Model/Abstract.php
@@ -96,7 +96,9 @@ class Core_Model_Abstract
     }
     public function delete()
     {
-        $this->getResource()->delete($this);
+        if($this->getId()){
+            $this->getResource()->delete($this);
+        }
         return $this;
     }
     // public function update()
